Corrections, commentaires et formattage

+++ Projet_Bibliotheque/src/java/com/robillard/bibliotheque/controlleur/SupprimerOuvrage.java
@@ -1,4 +1,6 @@
-
+//Supprime un ouvrage de la bd
+//L'utilisateur doit être connecté et être de type 2 (employé)
+//Le id soumis doit correspondre à un ouvrage dans la base de données
 package com.robillard.bibliotheque.controlleur;
 
 import com.mysql.jdbc.Connection;
@@ -6,7 +8,6 @@ import com.robillard.bibliotheque.modele.classes.Ouvrage;
 import com.robillard.bibliotheque.modele.dao.OuvrageDAO;
 import com.robillard.bibliotheque.util.Connexion;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.net.URLEncoder;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -16,42 +17,53 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class SupprimerOuvrage extends HttpServlet {
+public class SupprimerOuvrage extends HttpServlet
+{
 
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        try
+            throws ServletException, IOException
+    {
+        if (request.getSession().getAttribute("type") != null
+                && Integer.parseInt(request.getSession().getAttribute("type").toString()) >= 2)
         {
-            Class.forName(this.getServletContext().getInitParameter("piloteJDBC"));
-            Connexion.setUrl(this.getServletContext().getInitParameter("urlBd"));
-            Connection cnx = (Connection) Connexion.getInstance();
-            OuvrageDAO dao = new OuvrageDAO(cnx);
-            Ouvrage ouvrage = dao.read(request.getParameter("id"));
-            if (ouvrage != null)
+            try
             {
-                dao.delete(ouvrage);
-                String message = "L'ouvrage a " + URLEncoder.encode("é", "UTF-8") 
-                                + "t" + URLEncoder.encode("é", "UTF-8") + 
-                                " supprim" + URLEncoder.encode("é", "UTF-8") + 
-                                " avec succ" + URLEncoder.encode("è", "UTF-8") + "s";
-                response.sendRedirect("go?action=afficherGestionCatalogue&message="+message
-                                      +"&recherche="+request.getParameter("recherche")
-                                      +"&critere="+request.getParameter("critere"));
+                Class.forName(this.getServletContext().getInitParameter("piloteJDBC"));
+                Connexion.setUrl(this.getServletContext().getInitParameter("urlBd"));
+                Connection cnx = (Connection) Connexion.getInstance();
+                OuvrageDAO dao = new OuvrageDAO(cnx);
+                Ouvrage ouvrage = dao.read(request.getParameter("id"));
+                if (ouvrage != null)
+                {
+                    dao.delete(ouvrage);
+                    String message = "L'ouvrage a " + URLEncoder.encode("é", "UTF-8")
+                            + "t" + URLEncoder.encode("é", "UTF-8")
+                            + " supprim" + URLEncoder.encode("é", "UTF-8")
+                            + " avec succ" + URLEncoder.encode("è", "UTF-8") + "s";
+                    response.sendRedirect("go?action=afficherGestionCatalogue&message=" + message
+                            + "&recherche=" + request.getParameter("recherche")
+                            + "&critere=" + request.getParameter("critere"));
+                }
+                else
+                {
+                    RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
+                    r.forward(request, response);
+                }
             }
-            else
+            catch (Exception exp)
             {
+                Logger logger = Logger.getLogger("monLogger");
+                logger.log(Level.SEVERE, exp.getMessage());
+                String message = "Une erreur inattendue s'est produite. Veuillez"
+                        + " réessayer plus tard.";
+                request.setAttribute("erreurException", message);
                 RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
-                r.forward(request, response);  
+                r.forward(request, response);
             }
         }
-        catch (Exception exp)
+        else
         {
-            Logger logger = Logger.getLogger("monLogger");
-            logger.log(Level.SEVERE, exp.getMessage());
-            String message = "Une erreur inattendue s'est produite. Veuillez"
-                    + " réessayer plus tard.";
-            request.setAttribute("erreurException", message);
-            RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
+            RequestDispatcher r = this.getServletContext().getRequestDispatcher("/index.jsp");
             r.forward(request, response);
         }
     }
@@ -67,7 +79,8 @@ public class SupprimerOuvrage extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
+            throws ServletException, IOException
+    {
         processRequest(request, response);
     }
 
@@ -81,7 +94,8 @@ public class SupprimerOuvrage extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
+            throws ServletException, IOException
+    {
         processRequest(request, response);
     }
 
@@ -91,7 +105,8 @@ public class SupprimerOuvrage extends HttpServlet {
      * @return a String containing servlet description
      */
     @Override
-    public String getServletInfo() {
+    public String getServletInfo()
+    {
         return "Short description";
     }// </editor-fold>
